projects/personal_telegram_bot/requirements.txt
# Core Telegram Bot
python-telegram-bot==20.7

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# HTTP client (compatible with supabase)
httpx>=0.24.0,<0.25.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Voice processing and AI
openai==1.3.0
pydub==0.25.1
SpeechRecognition==3.10.0

# Optional: For advanced AI features
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2

# Legacy support
pathlib2==2.3.7
